% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTT_algorithmGreen.R
\name{FTT_algorithmGreen}
\alias{FTT_algorithmGreen}
\title{Algorithm to test for fixing taxon traceability}
\usage{
FTT_algorithmGreen(data, verbose = F, c = 4, n)
}
\arguments{
\item{data}{Data.table as constructed by create input. All possible c-tuples given the taxon set with status information (c-tuple as input available, c-tuple not in input = unresolved). In addition, all c (c-1)-tuples possible by each c-tuple are listed.}

\item{verbose}{Logical parameter if message should be printed, Default: F}

\item{c}{Parameter indicating the size of the c-tuples. E.g. if c=4, one uploads quadruples to be tested for fixing taxon traceability. Default: 4}

\item{n}{Number of all taxa in the analysis}
}
\value{
The same data.table is returned, with updated status, fixing taxon & round of resolvement.
}
\description{
Testing the input data stepwise for fixing taxa and check if the whole set can be resolved or not. The algorithm follows strictly the Mathematica template from Mareike Fischer by checking each green c-tuple if it can be used to solve a red one. There are two counters: first a simple counter adding up all solved c-tuples (green + newGreens). If this reaches the maximal number of c-tuples, the while loops stops. Second, I use a count to track the number of while and for loops used.
}
\details{
See publication and master thesis
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{copy}}
\code{\link[foreach]{foreach}}
}
