% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTT_algorithmRed_helpeR.R
\name{FTT_algorithmRed_helpeR}
\alias{FTT_algorithmRed_helpeR}
\title{Helper function for the \emph{red} algorithm for fixing taxon traceability.}
\usage{
FTT_algorithmRed_helpeR(data, roundnumber, verbose = F, c = 4, n)
}
\arguments{
\item{data}{Data.table as constructed by create input + count and fixing taxon. All possible c-tuples given the taxon set with status information (c-tuple as input available, c-tuple not in input = unresolved, c-tuples already solved = resolved). In addition, all c (c-1)-tuples possible by each c-tuple are listed. For resolved c-tuples, used fixing taxon and round of resolvement is listed.}

\item{roundnumber}{Index for round}

\item{verbose}{Logical parameter if message should be printed, Default: F}

\item{c}{Parameter indicating the size of the c-tuples. E.g. if c=4, one uploads quadruples to be tested for fixing taxon traceability. Default: 4}

\item{n}{Number of all taxa in the analysis}
}
\value{
The same data.table is returned, with updated status, fixing taxon & round
}
\description{
Split data to input and unresolved c-tuples and try for each unresolved one if a fixing taxon can be found.
}
\details{
See publication and master thesis
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[foreach]{foreach}}
\code{\link[data.table]{data.table-package}},\code{\link[data.table]{copy}},\code{\link[data.table]{rbindlist}}
}
