% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTT_createInput.R
\name{FTT_createInput}
\alias{FTT_createInput}
\title{Create Input for FixingTaxonTraceR Package}
\usage{
FTT_createInput(fn, sepSym, c = 4)
}
\arguments{
\item{fn}{Path to text file containing the taxon list. Taxa within can be numbers or characters. If the taxon lists have different length, missing entries will be filled with NA (if input is numeric) or "" (if input are characters).}

\item{sepSym}{Character or symbol used to separate the taxa, e.g. "_" or ","}

\item{c}{Parameter indicating the size of the c-tuples. E.g. if c=4, one uploads quadruples to be tested for fixing taxon traceability. Default: 4}
}
\value{
The output is a list containing the 5 data sets:
\itemize{
\item input_raw: data as given in the text file.
\item input_ctuples: interim result, data transformed into c-tuples only
\item input_ordered: data transformed to c-tuples, and taxa ID forced to numeric and ordered hierarchically
\item data: all possible c-tuples given the taxon set with status information (c-tuple as input available, c-tuple not in input = unresolved). In addition, all possible (c-1)-tuples by each c-tuple are listed
\item taxa: data table used for transformation, taxaID denotes the original input taxaID (as in input_raw & input_quadruples), NR is the ordered number of this taxon (as in input_ordered & data)
}
}
\description{
The function takes a text file with taxon lists (c-tuples or more) and transforms them to target format (only c-tuples, numbered, ordered)
}
\details{
\enumerate{
\item Get number of taxon lists (trees), unique taxa count, and transformation matrix
\item Split larger list into quadruples
\item Force numeric & ordered taxa (usind transformation matrix)
\item Get all possible quadruples given the number of unique input taxa
\item Get status information & return list
}
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{setDTthreads}},\code{\link[data.table]{fread}},\code{\link[data.table]{data.table-package}},\code{\link[data.table]{setorder}},\code{\link[data.table]{as.data.table}},\code{\link[data.table]{rbindlist}},\code{\link[data.table]{setattr}},\code{\link[data.table]{copy}}
\code{\link[foreach]{foreach}}
}
