% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTT_findNRC.R
\name{FTT_findNRC}
\alias{FTT_findNRC}
\title{Find no-rainbow-coloring for input quadruples}
\usage{
FTT_findNRC(data, verbose = F, verbose2 = T)
}
\arguments{
\item{data}{The input data is the list created by \emph{FTT_createInput}, both the data table \emph{taxa} and \emph{data} is used}

\item{verbose}{Logical parameter if all messages should be printed, Default: F}

\item{verbose2}{Logical parameter if final message should be printed, Default: T}
}
\value{
A list containing three sets:
\itemize{
\item the coloring per taxa
\item the colored quadruples
\item the result: \emph{fail} indicates decisiveness (no coloring was a no-rainbow 4-coloring), while \emph{NRC} indicates the first no-rainbow 4-coloring and the reason to call the set not-decisive
}
}
\description{
This function searches for a no-rainbow 4-coloring of the input quadruples. If there is one, then it represents one 4-way partition that is not covered by the input quadruples - hence the qaudruple set is not phylogenetic decisive.

\strong{IMPORTANT NOTE}: This function is only meaningful for c=4.
}
\details{
This algorithm was adapted from Algorithm 2 (No-Rainbow 4-Coloring of H) of \href{https://ieeexplore.ieee.org/document/9616390}{Ghazaleh Parvini, Katherine Braught, and David Fernandez-Baca} (DOI: 10.1109/TCBB.2021.3128381)
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{copy}}
}
