% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FTT_initialCheck.R
\name{FTT_initialCheck}
\alias{FTT_initialCheck}
\title{Perform some initial checks}
\usage{
FTT_initialCheck(data, c = 4, n, verbose = F)
}
\arguments{
\item{data}{Input data created by FTT_createInput()}

\item{c}{Parameter indicating the size of the c-tuples. E.g. if c=4, one uploads quadruples to be tested for fixing taxon traceability. Default: 4}

\item{n}{Number of all taxa in the analysis}

\item{verbose}{Logical parameter if message should be printed, Default: F}
}
\value{
Results of the three checks
}
\description{
Three checks: First, the bounds of the FTT algorithm are checked. If the number of c-tuples is above the upper or below the lower bound, then the FTT algorithm does not to be used as the result is fix. Second, three properties of phylogenetic decisiveness are tested for c=4: are all triples covered? are triples with simple covered covered by different quadruples? are all tuples sufficiently covered?
}
\details{
See publication and master thesis
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[data.table]{melt.data.table}},\code{\link[data.table]{copy}}
}
